third_party/WebKit/LayoutTests/crypto/subtle/aes-key-algorithm-mismatch-expected.txt

Call encrypt using the wrong AES key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS key.algorithm.name is "AES-GCM"
error is: InvalidAccessError: key.algorithm does not match that of operation
PASS successfullyParsed is true

TEST COMPLETE
